requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.5
distlib==0.3.9
Django==5.2.11
django-allauth==65.14.3
django-cors-headers==4.9.0
django-filter==25.2
django-js-asset==3.1.2
django-mptt==0.18.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
filelock==3.16.1
gunicorn==25.1.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==12.1.1
platformdirs==4.3.6
pluggy==1.6.0
psycopg2-binary==2.9.11
pycparser==3.0
Pygments==2.19.2
PyJWT==2.11.0
PyMySQL==1.1.2
pytest==9.0.2
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
virtualenv==20.28.1
whitenoise==6.12.0
